Fixed problem with __FILE__ and RubyGems
When RubyGems executes a cli on windows, it wraps it in a
runner script.  Because of this, __FILE__ is not $0, and the program
does not work.

To rectify this, I moved the program into the Lib directory where
it really belongs and created my own minimal runner scripts in bin
to execute it.

I also renamed the App classes to Tangle and Lmw respectively.
Though Tangle is defined in lmt.ruby, this is because of name reuse
problems.

Also, install now depends on build so that the tangle will be built
before being installed

Finally, the classes are now in the Lmt Module.