Enabled conditional output
Because conditional output and evaling extensions are related,
it is important for them to run concurrently.  Because of this,
evaling the extensions is now processed inline using side effects
rather than by partitioning the input collection.

In order to keep the algorithms seperate, the conditional processing
is implemented as a class which contains the state of the algorithm
The two algorithms are then weaved together using lazy evaluation on
the filtering.